refactored lookup
